# api-docs/API.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'Online-LMS APIs'
  description: 'List of REST APIs Documentation for Online Learning Management System'
  contact:
    name: Palden Sherpa
    email: sherpalden369@gmail.com
servers: 
  - url: http://localhost:5000
    description: Local Server
    
  - url: https://aiex-olms.herokuapp.com
    description: Heroku Server
paths:
...
  /api/admin/dictionary:
    post:
      tags:
        - Dictionary APIs - Admin
      description: Create a dictionary post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content: 
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                keyword:
                  type: string
                  example: Frame
                description:
                  type: string
                  example: It is a structural design of building.
                youtubeLinks:
                  type: array
                  example: ["https://youtu.be/QNInfBuBcOA"]
                relatedWords:
                  type: array
                  example: ["beam", "building", "structure"]
      responses:
        200:
          description: Response on successful creation of dictionary post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: dictionary Post Creation Successful
  /api/admin/dictionary/bulk-upload:
    post:
      tags:
        - Dictionary APIs - Admin
      description: Bulk upload dictionary
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content: 
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  
      responses:
        200:
          description: Response on successful bulk upload of dictionary
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: dictionary bulk upload Successful
  /api/admin/dictionary/{dictionaryID}:
    put:
      tags:
        - Dictionary APIs - Admin
      description: Update a dictionary post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: dictionaryID
      requestBody:
        required: true
        content: 
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                imagesToBeDeleted:
                  type: array
                  example: ["img-1.jpg", "img-2.jpg"]
                keyword:
                  type: string
                  example: Frame
                description:
                  type: string
                  example: It is a structural design of building.
                youtubeLinks:
                  type: array
                  example: ["https://youtu.be/QNInfBuBcOA"]
                relatedWords:
                  type: array
                  example: ["beam", "building", "structure"]
      responses:
        200:
          description: Response on successful Update of dictionary post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: dictionary Post Update Successful

    delete:
      tags:
        - Dictionary APIs - Admin
      description: Delete a dictionary post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: dictionaryID
      responses:
        200:
          description: Response on successful Delete of dictionary post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: dictionary Post Delete Successful
  /api/admin/dictionary/search:
    get:
      tags:
        - Dictionary APIs - Admin
      description: Search a dictionary post
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: searchKey
        - in: query
          name: skips
        - in: query
          name: limit
     
      responses:
        200:
          description: Response on successful Search of dictionary post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: dictionary Post Search Successful


#******
#******
#******
#******
#******
#******


#******
#******
#******
#******
#******
#******
...
  /api/admin/course/add-main-category:
    post:
      tags:
        - Course APIs - Admin
      description: Add main course category
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              type: object
              properties:
                categoryName:
                  type: string
                  example: Science and myths
                  
      responses:
        200:
          description: Response on successful addition of main course category
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Course main category addition successful"
  /api/admin/course/add-sub-category:
    post:
      tags:
        - Course APIs - Admin
      description: Add sub category for course
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              type: object
              properties:
                categoryName:
                  type: string
                  example: Animal
                parentID:
                  type: string
                  example: adfadf7987sdfs98798sdf
                  
      responses:
        200:
          description: Response on successful addition of sub category
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Course sub category addition successful"
  /api/admin/course/rename-category:
    put:
      tags:
        - Course APIs - Admin
      description: Rename a category for course
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              type: object
              properties:
                newName:
                  type: string
                  example: Animal
                categoryID:
                  type: string
                  example: adfadf7987sdfs98798sdf
                  
      responses:
        200:
          description: Response on successful rename of a category
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Category rename successful"
  /api/admin/course/all-categories:
    get:
      tags:
        - Course APIs - Admin
      description: Get all category list of Books
      security:
        - bearerAuth: []
      responses:
        200:
          description: Response on successful retrieval of all category list
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "All category list retrieval successful"                 
  /api/admin/course:
    post:
      tags:
        - Course APIs - Admin
      description: Create a course post
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content: 
          multipart/form-data:
            schema:
              type: object
              properties:
                thumbnail:
                  type: string
                  format: binary
                files:
                  type: array
                  items:
                     type: string
                     format: binary
                title:
                  type: string
                  example: Earthquake Resistent Building Design
                authors:
                  type: array
                  items:
                    type: object
                    properties:
                      firstName:
                        type: string
                        example: Palden
                      lastName:
                        type: string
                        example: Sherpa
                level:
                  type: string
                  example: beginner
                categoryID:
                  type: string
                  example: 5454sdfasd54545sdf
                description:
                  type: object
                  example: ck editor object here...
               
      responses:
        200:
          description: Response on successful creation of course post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: course Post Creation Successful
  /api/admin/course/update-details/{courseID}:
    put:
      tags:
        - Course APIs - Admin
      description: Update a course details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseID
      requestBody:
        required: true
        content: 
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Earthquake Resistent Building Design
                authors:
                  type: array
                  items:
                    type: object
                    properties:
                      firstName:
                        type: string
                        example: Palden
                      lastName:
                        type: string
                        example: Sherpa
                level:
                  type: string
                  example: beginner
                categoryID:
                  type: string
                  example: 5454sdfasd54545sdf
                description:
                  type: object
                  example: ck editor object here...
      responses:
        200:
          description: Response on successful update of course post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: course Post Update Successful
  /api/admin/course/update-files/{courseID}:
    put:
      tags:
        - Course APIs - Admin
      description: Update a course files
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseID
      requestBody:
        required: true
        content: 
          multipart/form-data:
            schema:
              type: object
              properties:
                thumbnail:
                  type: string
                  format: binary
                files:
                  type: array
                  items:
                     type: string
                     format: binary
                filesToBeDeleted:
                  type: array
                  example: ["file1.pdf", "file2.pdf"]
                
      responses:
        200:
          description: Response on successful update of course files.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: course files Update Successful
  /api/admin/course/{courseID}:
    get:
      tags:
        - Course APIs - Admin
      description: Get details of a course
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseID
      responses:
        200:
          description: Response on successful retrieval of course details.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: course details retrieval Successful

    delete:
      tags:
        - Course APIs - Admin
      description: Delete a course post
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: courseID
      responses:
        200:
          description: Response on successful deletion of course post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: course Post Delete Successful
  /api/admin/course/all-courses/{categoryID}:
    get:
      tags:
        - Course APIs - Admin
      description: Get all courses of particular category
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: categoryID
        - in: query
          name: skips
        - in: query
          name: limit
      responses:
        200:
          description: Response on successful retrieval of courses of particular categoryID.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: courses of particular categoryID retrieval successful
  /api/admin/course/search:
    get:
      tags:
        - Course APIs - Admin
      description: Search a course post
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: searchKey
        - in: query
          name: skips
        - in: query
          name: limit
     
      responses:
        200:
          description: Response on successful Search of course post.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: course Post Search Successful


#******
#******
#******
#******
#******
#******
components:
  securitySchemes:
    bearerAuth:           
      type: http
      scheme: bearer
      bearerFormat: JWT  